Move services_running_here, add autospec=True
