+++ tests/test_firewall_update.py
@@ -67,66 +67,20 @@ def test_parse_args_default_cron():
     assert not args.verbose
 
 
-@pytest.yield_fixture
-def mock_get_running_mesos_docker_containers():
-    with mock.patch.object(
-        firewall_update, 'get_running_mesos_docker_containers', autospec=True,
-        return_value=[
-            {
-                'HostConfig': {'NetworkMode': 'bridge'},
-                'Labels': {
-                    'paasta_service': 'myservice',
-                    'paasta_instance': 'hassecurity',
-                },
-                'NetworkSettings': {
-                    'Networks': {
-                        'bridge': {'MacAddress': '02:42:a9:fe:00:0a'},
-                    },
-                },
-            },
-            {
-                'HostConfig': {'NetworkMode': 'bridge'},
-                'Labels': {
-                    'paasta_service': 'myservice',
-                    'paasta_instance': 'chronoswithsecurity',
-                },
-                'NetworkSettings': {
-                    'Networks': {
-                        'bridge': {'MacAddress': '02:42:a9:fe:00:0b'},
-                    },
-                },
-            },
-            # host networking
-            {
-                'HostConfig': {'NetworkMode': 'host'},
-                'Labels': {
-                    'paasta_service': 'myservice',
-                    'paasta_instance': 'batch',
-                },
-            },
-            # no labels
-            {
-                'HostConfig': {'NetworkMode': 'bridge'},
-                'Labels': {},
-            },
-        ],
-    ):
-        yield
-
-
-@pytest.mark.usefixtures('mock_get_running_mesos_docker_containers')
-def test_services_running_here():
-    assert tuple(firewall_update.services_running_here()) == (
+@mock.patch.object(
+    firewall_update, 'load_system_paasta_config', autospec=True,
+    return_value=mock.Mock(**{
+        'get_cluster.return_value': 'mycluster',
+    }),
+)
+@mock.patch.object(
+    firewall, 'services_running_here', autospec=True,
+    return_value=(
         ('myservice', 'hassecurity', '02:42:a9:fe:00:0a'),
         ('myservice', 'chronoswithsecurity', '02:42:a9:fe:00:0b'),
-    )
-
-
-@pytest.mark.usefixtures('mock_get_running_mesos_docker_containers')
-@mock.patch.object(firewall_update, 'load_system_paasta_config', return_value=mock.Mock(**{
-    'get_cluster.return_value': 'mycluster',
-}))
-def test_smartstack_dependencies_of_running_firewalled_services(_, tmpdir):
+    ),
+)
+def test_smartstack_dependencies_of_running_firewalled_services(_, __, tmpdir):
     soa_dir = tmpdir.mkdir('yelpsoa')
     myservice_dir = soa_dir.mkdir('myservice')
